Upgrade to the latest SvelteKit
