+++ sveltekit_mdsvex/src/lib/util.ts
@@ -1,1 +1,1 @@
-export const slugFromPath = (path) => path.match(/([\w-]+)\.(svelte\.md|md|svx)/i)?.[1] ?? null;
+export const slugFromPath = (path) => path.match(/(.+\/)*(.+)\/(.+)\.(svelte\.md|md|svx)/i)?.[2] ?? null;
