+++ sveltekit_mdsvex/src/routes/posts/[slug].json/+server.ts
@@ -1,10 +1,11 @@
+import { json } from '@sveltejs/kit';
 import { slugFromPath } from '$lib/util';
 
 /**
  * @type {import('@sveltejs/kit').RequestHandler}
  */
-export async function get({ params }) {
-    const modules = import.meta.glob(`./*.{md,svx,svelte.md}`);
+export async function GET({ params }) {
+    const modules = import.meta.glob(`../**/*.{md,svx,svelte.md}`);
 
     let match;
     for (const [path, resolver] of Object.entries(modules)) {
@@ -15,14 +16,9 @@ export async function get({ params }) {
     }
 
     if (!match) {
-        return {
-            status: 404
-        };
+        return new Response(undefined, { status: 404 });
     }
 
     const post = await match[1]();
-
-    return {
-        body: post.metadata
-    };
+    return json(post.metadata);
 }
