+++ sveltekit_mdsvex/src/routes/posts/post2/+page.svx
