+++ sveltekit_mdsvex/svelte.config.js
@@ -10,13 +10,7 @@ const config = {
 		}),
 	],
 	kit: {
-		adapter: adapter(),
-		// hydrate the <div id="svelte"> element in src/app.html
-		target: '#svelte',
-		prerender: {
-			onError: 'continue'
-		},
-		trailingSlash: 'always'
+		adapter: adapter()
 	}
 };
